-cLanguage classe casting long to int

+++ app/src/main/java/com/example/chbla/ba_eresamont/Classes/CLanguageID.java
@@ -14,7 +14,7 @@ public class CLanguageID {
  //check the array of pages-lang which item of array has the number
     public int GetLanguageID(Pages pages, int inputlang){
         for(int i=0;i<pages.getPages_lang().size();i++){
-            if (pages.getPages_lang().get(i).getLanguage()==inputlang)
+            if ((int)pages.getPages_lang().get(i).getLanguage()==inputlang)
                 mlang=i;
         }
         return mlang;
